+++ src/main/java/com/joaocarlos/seplag/repositories/ServidorEfetivoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServidorEfetivoRepository extends JpaRepository<ServidorEfetivo,Long> {
+public interface ServidorEfetivoRepository extends JpaRepository<ServidorEfetivo,Integer> {
 }
